Normalized Schema

+++ publisher_table_db2.sql
@@ -1,5 +1,6 @@
-CREATE TABLE Publishers (
+CREATE TABLE Publishers
+(
 	publisher_id INT IDENTITY(0, 1) NOT NULL PRIMARY KEY,
 	publisher_name VARCHAR(60),
-	publisher_revenue FLOAT
+	publisher_revenue DECIMAL(10, 2)
 );
